Release version 0.3.0 of the wasm_val crate :
 - Add the ability to send multiple values to javascript
 - Add the ability to call a function directly
 - Add the ability to call functions with multiple args
 - Rename previous call to call_method
 - Other arious refactoring

+++ wasm_val/src/js_ser/js_val.rs
@@ -6,7 +6,7 @@ use super::{JsSerializable};
 use byteorder::{LittleEndian, WriteBytesExt,};
 
 impl JsSerializable for JsValue {
-    fn size(&self) -> u8 {
+    fn size(&self) -> u32 {
         match &self.val {
             Val::None => 1,
             Val::Boolean(b) => b.size(),
